- SCR(s) Resolved: - Description: Clean up codes, remove all useless varibles ,casting and deprecated method - Regression (Yes / No): No - Code Owner: Yueqian Wang - Code Reviewers: Qiangsheng Wang - Tests: Manual test and JUnit test - Test Automated (Yes/No, if "No" then explain why):Yes - Case Entries Resolved: CVS/Home - Notes to Developers: - Notes to QA: - Notes to Documentation: - Notes to Configuration Management: - Notes to Support: - Notes to Product Marketing:

+++ UI/org.eclipse.birt.report.designer.ui/src/org/eclipse/birt/report/designer/internal/ui/editors/rulers/EditorRulerComposite.java
@@ -436,17 +436,17 @@ public class EditorRulerComposite extends Composite
 
 	private int getLeftMargin( )
 	{
-		return (int) ( getMasterPageInsets( getMasterPageHandle( ) ).left );
+		return ( getMasterPageInsets( getMasterPageHandle( ) ).left );
 	}
 
 	private int getRightMargin( )
 	{
-		return (int) ( getMasterPageInsets( getMasterPageHandle( ) ).right );
+		return ( getMasterPageInsets( getMasterPageHandle( ) ).right );
 	}
 
 	private int getBottomMargin( )
 	{
-		return (int) ( getMasterPageInsets( getMasterPageHandle( ) ).bottom );
+		return ( getMasterPageInsets( getMasterPageHandle( ) ).bottom );
 	}
 
 	private double getZoom( )
@@ -467,7 +467,7 @@ public class EditorRulerComposite extends Composite
 
 	private int getTopMargin( )
 	{
-		return (int) ( getMasterPageInsets( getMasterPageHandle( ) ).top );
+		return ( getMasterPageInsets( getMasterPageHandle( ) ).top );
 	}
 
 	private GraphicalViewer getRulerContainer( int orientation )
